Fix for restorFunciontions/moveFieldState
Any existing field in a state must have all 3 callbacks,
without atleast one of them a field is broken and a form will not work correctly.
The callbacks are created at the moment of a field's registration,
and they are created only when the field is not exist, so we must
not leave any field with a missing callback.

https://github.com/final-form/final-form/blob/master/src/FinalForm.js#L800-L821